chore: Rename LoginPage to SignInPage and adds scroll to it
