+++ lib/src/routes/routes.dart
@@ -3,17 +3,17 @@ import 'package:delivery_app/src/features/presentation/collections_page/view/col
 import 'package:delivery_app/src/features/presentation/filter_page/view/filter_page.dart';
 import 'package:delivery_app/src/features/presentation/forgot_password_page/view/forgot_password_page.dart';
 import 'package:delivery_app/src/features/presentation/home_page/home_page.dart';
-import 'package:delivery_app/src/features/presentation/login_page/view/login_page.dart';
 import 'package:delivery_app/src/features/presentation/place_detail_page/view/place_detail_page.dart';
 import 'package:delivery_app/src/features/presentation/profile_detail_page/view/profile_detail_page.dart';
 import 'package:delivery_app/src/features/presentation/search_page/view/search_page.dart';
+import 'package:delivery_app/src/features/presentation/sign_in_page/view/sign_in_page.dart';
 import 'package:delivery_app/src/features/presentation/sign_up_page/view/sign_up_page.dart';
 import 'package:delivery_app/src/features/presentation/welcome_page/view/welcome_page.dart';
 import 'package:flutter/widgets.dart';
 
 final routes = <String, WidgetBuilder>{
   'welcome': (BuildContext context) => const WelcomePage(),
-  'login': (BuildContext context) => const LoginPage(),
+  'sign-in': (BuildContext context) => const SignInPage(),
   'forgot-password': (BuildContext context) => const ForgotPasswordPage(),
   'sign-up': (BuildContext context) => SingUpPage(),
   'home': (BuildContext context) => const HomePage(),
